Small interface additions
Add `operator*` as an additional spelling of `get`. Revise the iterator
types to be raw pointers, because raw pointers are guaranteed to be
past-the-end incrementable, and the behavior of custom pointers is just
whatever anyhow.

Add a `from_raw` function to create a box from a raw pointer, although
this does not support custom allocators due to correct argument
deduction being impossible (at least I think so; if not, then it still
surpasses my current level of template skills).

As always, add a few tests for the added pieces.