Add `m_has_value` flag and `erase` operation
As it turns out, managing state solely through the stored pointer is
impossible if you want to keep memory around. Thus, we add a "has a
value" flag and refactor all methods to use it instead of a null pointer
check accordingly.

Also, we now have an `erase` operation that mimics the behavior of
`std::vector::erase`, i.e. it destroys the stored object (if any) but
keeps the memory around.